backend/pyproject.toml
[project]
name = "Horizon-News"
version = "0.1.0"
description = "An Internation news website"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi[all]>=0.128.8",
    "pwdlib[argon2]>=0.3.0",
    "pyjwt>=2.11.0",
    "python-multipart>=0.0.22",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.46",
]


[dependency-groups]
runtime = [
    "fastapi",
    "uvicorn[standard]",
    "sqlalchemy",
    "pydantic-settings",
    "pyjwt",
    "pwdlib",
    "httpx",
    "python-multipart",
]

dev = [
    "fastapi>=0.133.1",
    "ruff>=0.15.1",
    "sqlalchemy>=2.0.48",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["src/test"]
